requirements.txt
asgiref==3.9.1
Django==5.2.5
django-cors-headers==4.7.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
drf-yasg==1.21.10
gunicorn==23.0.0
inflection==0.5.1
packaging==25.0
psycopg2-binary==2.9.10
PyJWT==2.10.1
python-dotenv==1.1.1
pytz==2025.2
PyYAML==6.0.2
sqlparse==0.5.3
typing_extensions==4.15.0
uritemplate==4.2.0
whitenoise==6.9.0
